requirements.txt
# Core dependencies
requests==2.31.0
google-generativeai==0.8.3
python-dotenv==1.0.0
schedule==1.2.1

# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# Video downloading
yt-dlp==2024.12.6
